Don't query non-existent queues (#770)
