+++ test/Microsoft.Azure.Webjobs.Extensions.Storage.UnitTests/Blobs/AsyncResultExtensions.cs
