+++ test/Microsoft.Azure.Webjobs.Extensions.Storage.UnitTests/Blobs/Bindings/MockOfCloudBlobStreamExtensions.cs
