+++ test/Microsoft.Azure.Webjobs.Extensions.Storage.UnitTests/Blobs/Bindings/ReturnsThrowsOfCloudBlobStreamICancellableAsyncResultExtensions.cs
