+++ test/Microsoft.Azure.Webjobs.Extensions.Storage.UnitTests/Blobs/BlobTriggerTests.cs
