+++ test/Microsoft.Azure.Webjobs.Extensions.Storage.UnitTests/Blobs/CompletedAsyncResult.cs
