+++ test/Microsoft.Azure.Webjobs.Extensions.Storage.UnitTests/Blobs/CompletingAsyncResult.cs
