+++ test/Microsoft.Azure.Webjobs.Extensions.Storage.UnitTests/Blobs/Listeners/BlobLogListenerTests.cs
