+++ test/Microsoft.Azure.Webjobs.Extensions.Storage.UnitTests/Blobs/MockOfStreamExtensions.cs
