+++ test/Microsoft.Azure.Webjobs.Extensions.Storage.UnitTests/Blobs/SetupOfStreamIAsyncResultExtensions.cs
