+++ test/Microsoft.Azure.Webjobs.Extensions.Storage.UnitTests/Blobs/WatchableReadStreamTests.cs
