+++ test/Microsoft.Azure.Webjobs.Extensions.Storage.UnitTests/Queues/FakeQueueConfiguration.cs
