+++ test/Microsoft.Azure.Webjobs.Extensions.Storage.UnitTests/Queues/HostStopTests.cs
